feat: Draft. New route loaders. Refactor `ApiLoader`
Refs #3539
